+++ src/test/java/examples/sharding/TableCodesDynamicSqlSupport.java
@@ -1,0 +1,35 @@
+/*
+ *    Copyright 2016-2022 the original author or authors.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package examples.sharding;
+
+import org.mybatis.dynamic.sql.AliasableSqlTable;
+import org.mybatis.dynamic.sql.SqlColumn;
+
+import java.sql.JDBCType;
+
+public final class TableCodesDynamicSqlSupport {
+    public static final TableCodes tableCodes = new TableCodes();
+    public static final SqlColumn<Integer> id = tableCodes.id;
+    public static final SqlColumn<String> description = tableCodes.description;
+
+    public static final class TableCodes extends AliasableSqlTable<TableCodes> {
+        public final SqlColumn<Integer> id = column("id", JDBCType.INTEGER);
+        public final SqlColumn<String> description = column("description", JDBCType.VARCHAR);
+        public TableCodes() {
+            super("tableCodes_even", TableCodes::new);
+        }
+    }
+}