+++ src/test/resources/examples/sharding/ShardingDB.sql
@@ -1,0 +1,30 @@
+--
+--    Copyright 2016-2022 the original author or authors.
+--
+--    Licensed under the Apache License, Version 2.0 (the "License");
+--    you may not use this file except in compliance with the License.
+--    You may obtain a copy of the License at
+--
+--       https://www.apache.org/licenses/LICENSE-2.0
+--
+--    Unless required by applicable law or agreed to in writing, software
+--    distributed under the License is distributed on an "AS IS" BASIS,
+--    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+--    See the License for the specific language governing permissions and
+--    limitations under the License.
+--
+
+drop table tableCodes_even if exists;
+drop table tableCodes_odd if exists;
+
+create table tableCodes_even (
+    id int not null,
+    description varchar(30) not null,
+    primary key (id)
+);
+
+create table tableCodes_odd (
+    id int not null,
+    description varchar(30) not null,
+    primary key (id)
+);